Time: 0 ms (100.00%), Space: 5.9 MB (88.42%) - LeetHub

+++ 171-excel-sheet-column-number/171-excel-sheet-column-number.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int titleToNumber(string columnTitle) {
-        long long num=0;
+        int num=0;
         reverse(columnTitle.begin(), columnTitle.end());
-    for (long long i = 0; i < columnTitle.length(); i++) {
-        long long temp = columnTitle[i] - '@';
+    for (int i = 0; i < columnTitle.length(); i++) {
+        int temp = columnTitle[i] - '@';
             int j = i;
             while (j--)
                 temp *= 26;
